check configs
- New process for creating configs.yaml
- add features in check_config()
